Add TTL cache to redfish discover method
The `redfish.discover_ssdp()` method takes a very long time to query the
redfish services. This causes the scrape timeout to be exceeded for
Prometheus. In order to solve this, a TTL cache has been added. A
logging message has also been added for debugging.

Additionally, the `redfish.redfish_client` call to create a redfish
object has been provided with the `timeout` and `max_retry` arguments in
order to not have the query take too long in case of incorrect redfish
credentials.