Add TTL cache to redfish discover method (#34)
The `redfish.discover_ssdp()` method takes a very long time to query the
redfish services. This causes the scrape timeout to be exceeded for
Prometheus. In order to solve this, a TTL cache has been added. A
logging message has also been added for debugging.

Additionally, the `redfish.redfish_client` call to create a redfish
object has been provided with the `timeout` and `max_retry` arguments in
order to not have the query take too long in case of incorrect redfish
credentials.

redfish_client config and discover cache parameters can be
modified through the prometheus-hardware-exporter CLI args.
The default values for these configs are defined in `config.py`.

Some other helpful improvements:
- `RedfishHelper` class now directly takes an instance of `Config` instead of 
sending the parameters individually.
- Use context manager for redish login object to automatically logout at the end of the context.
- Split sensor data methods to be more readable.
- Add unit tests