Refactor redfish discover method
If the discover method is in the RedfishHelper class, the instance is deleted at the end of the context manager execution.
This means that the cache is not being used and we have a huge lag each time the metrics are being scraped.

Since the discover method doesn't need a redfish session, we can move it out into its own class so that it can keep the lifetime of the RedfishCollector class and maintain the cache.
We also don't need to worry about any session login/logout.