fix(redfish): Make redfish storage name in URI dynamic.
The redfish storage name in the URI was initially hardcoded to "Storage".
On some servers which did not conform to the schema specification, this
name was provided differently, eg: Storages.

This change aims to find the storage uri name dynamically while fetching
the storage controller and storage drive data.

Fixes #91, #108